+++ tests/t20052/t20052.cc
@@ -1,0 +1,105 @@
+#include <algorithm>
+#include <functional>
+#include <memory>
+#include <optional>
+#include <utility>
+#include <vector>
+
+namespace clanguml {
+namespace t20052 {
+struct A {
+    void a() { aa(); }
+
+    void aa() { aaa(); }
+
+    void aaa() { }
+};
+
+struct B {
+    void b() { bb(); }
+
+    void bb() { bbb(); }
+
+    void bbb() { }
+
+    void eb() { }
+};
+
+struct C {
+    void c() { cc(); }
+
+    void cc() { ccc(); }
+
+    void ccc() { }
+};
+
+struct D {
+    int add5(int arg) const { return arg + 5; }
+};
+
+class E {
+    std::optional<std::shared_ptr<B>> maybe_b;
+    std::shared_ptr<A> a;
+
+public:
+    template <typename F> void setup(F &&f) { f(maybe_b); }
+};
+
+template <typename F> struct R {
+    R(F &&f)
+        : f_{std::move(f)}
+    {
+    }
+
+    void r() { f_(); }
+
+    F f_;
+};
+
+void tmain()
+{
+    A a;
+    B b;
+    C c;
+
+    // The activity shouldn't be marked at the lambda definition, but
+    // wherever it is actually called...
+    auto alambda = [&a, &b]() {
+        a.a();
+        b.b();
+    };
+
+    // ...like here
+    alambda();
+
+    // There should be no call to B in the sequence diagram as the blambda
+    // is never called
+    [[maybe_unused]] auto blambda = [&b]() { b.b(); };
+
+    // Nested lambdas should also work
+    auto clambda = [alambda, &c]() {
+        c.c();
+        alambda();
+    };
+    clambda();
+
+    R r{[&c]() { c.c(); }};
+
+    r.r();
+
+    D d;
+
+    std::vector<int> ints{0, 1, 2, 3, 4};
+    std::transform(ints.begin(), ints.end(), ints.begin(),
+        [&d](auto i) { return d.add5(i); });
+
+    // TODO: Fix naming function call arguments which are lambdas
+    //    E e;
+    //
+    //    e.setup([](auto &&arg) mutable {
+    //        // We cannot know here what 'arg' might be
+    //        arg.value()->eb();
+    //    });
+}
+}
+}